requirements.txt
requests
psycopg2-binary
psycopg2
pandas
pandas_gbq
python-dotenv
celery
redis
gspread
flower
celery-singleton
sqlalchemy
# pyspark==3.1.2 
airflow-code-editor
# apache-airflow-providers-apache-spark
# apache-airflow-providers-airbyte[http]